# perf: fully cached shapetracker methods