Not necessarily for merge, mainly to show @chaosagent.

Kept the external shapetracker interfaces, this will be a bit faster still (and way less lines) if those are rewritten in lazy etc.

LLama:

Before
```
codegen         mean runtime:  238.50ms, runs:   293.03,  224.61,  220.70,  243.19,  222.52,  229.64,  224.70,  224.93,  223.50,  278.18
methodcache     mean runtime:  212.00ms, runs:   189.93,  210.70,  216.78,  207.13,  207.24,  235.79,  232.11,  211.25,  203.36,  205.74
profile         mean runtime: 1082.35ms, runs:  1013.85, 1055.22, 1052.08, 1030.62, 1129.18, 1271.14, 1087.14, 1049.98, 1069.15, 1065.08
```


After
```
codegen         mean runtime:  211.45ms, runs:   267.03,  200.79,  203.35,  192.95,  193.02,  194.79,  262.67,  202.87,  197.08,  199.89
methodcache     mean runtime:  185.16ms, runs:   162.61,  181.66,  190.57,  176.26,  176.46,  178.97,  189.02,  240.30,  180.14,  175.58
profile         mean runtime:  860.25ms, runs:   789.98,  791.11,  792.08,  841.22,  948.47,  920.92,  902.69,  857.30,  887.20,  871.53
```


CIFAR
Before
```
 PYTHONPATH=. STEPS=5 WINO=1 python3.11 -O examples/hlb_cifar10.py
  0 34880.25 ms run, 34880.09 ms python,    0.16 ms CL, 1187.23 loss, 0.002563 LR, 0.84 GB used,     13.61 GFLOPS
  1 8163.48 ms run, 8163.33 ms python,    0.15 ms CL, 1187.59 loss, 0.002577 LR, 4.39 GB used,     57.37 GFLOPS
  2  218.71 ms run,    8.44 ms python,  210.27 ms CL, 2760.15 loss, 0.001741 LR, 4.39 GB used,   2141.32 GFLOPS
  3  222.95 ms run,    8.38 ms python,  214.57 ms CL, 2844.72 loss, 0.000906 LR, 4.39 GB used,   2100.54 GFLOPS
  4  228.68 ms run,    8.68 ms python,  220.00 ms CL, 2118.59 loss, 0.000070 LR, 4.39 GB used,   2047.96 GFLOPS
```

After
```
  0 28633.72 ms run, 28633.58 ms python,    0.14 ms CL, 1187.23 loss, 0.002563 LR, 0.84 GB used,     16.58 GFLOPS
  1 7006.18 ms run, 7006.04 ms python,    0.14 ms CL, 1187.59 loss, 0.002577 LR, 4.39 GB used,     66.84 GFLOPS
  2  218.80 ms run,    8.20 ms python,  210.59 ms CL, 2760.15 loss, 0.001741 LR, 4.39 GB used,   2140.45 GFLOPS
  3  224.61 ms run,   12.93 ms python,  211.67 ms CL, 2844.72 loss, 0.000906 LR, 4.39 GB used,   2085.09 GFLOPS
  4  227.51 ms run,   11.21 ms python,  216.30 ms CL, 2118.59 loss, 0.000070 LR, 4.39 GB used,   2058.45 GFLOPS
```