@roelofvandijk 2023-09-10
@geohot This is ~20% for llama and winograd, and probably most other things. Only caching, no functional changes. The diff is shot in this MR - should I split it up into parts? I can reduce the linecount delta as well, but that requires changing the Shapetracker interface (&tests), which I'd prefer to do in a separate MR. A next simplification would be to make the Shapetracker a namedtuple, but that's also nice as a separate MR.