@geohot 2023-09-10
Does this increase/reduce complexity/readability?