@geohot 2023-09-10
My proposal would be to split this into two files, view.py and shapetracker.py, with an eye toward deleting shapetracker.py