@roelofvandijk 2023-09-10
> Does this increase readability? 

Yes, I think so. By only passing what is needed to each function, it becomes clear that the whole Shapetracker class is just a pretty interface. Also, all caches work better.

E.g., these six methods. 
```
  def reshape(self, new_shape): 
    if self.views[-1].shape == new_shape: return self
    self.views = reshape(self.views, new_shape)
    return self
  def expand(self, new_shape):
    self.views = self.views[:-1] + (expand(self.views[-1], new_shape),)
    return self
  def shrink(self, arg):
    self.views = self.views[:-1] + (shrink(self.views[-1], arg),)
    return self
  def stride(self, mul):
    self.views = self.views[:-1] + (stride(self.views[-1], mul),)
    return self
  def permute(self, axis):
    self.views = self.views[:-1] + (permute(self.views[-1], axis),)
    return self
  def pad(self, arg):
    self.views = self.views[:-1] + (pad(self.views[-1], arg),)
    return self
```

> Layer of indirection

The extra layer of indirection is only needed to keep the interface of the ShapeTracker class the same (tests & usages in lazy), it's also a matter of taste, e.g. which of these you prefer:
`view.expr_idxs(idxs)`
`expr_idxs(view, idxs)`

The second one is way better cacheable. You seem to prefer the first one. If the speed matters, extra layer it is.

Is it focused on readability? No. I just wanted to create an example of how to fully cache all functions here without touching any other files.

> Removing shapetracker & creating view.py: 

That's absolutely the way to go. I have a working branch where each buffer only has a view of tuples (instead of a shapetracker). The five methods (except for `reshape`) that modify a single view could move to the view class right away, but there is something to be said for keeping the set of 6 together.

A possible upgrade path would be (this is what I did):

* this (making the methods on the views transparent)
* `Shapetracker(NamedTuple)` -> becomes immutable, needs to be re-created on change, which forces some changes in `lazy` etc.
* remove the ShapeTracker wrapper, each buffer has tuples of views:
`def create_lazybuffer(device:str, st:Tuple[View], optype:OpType, ... `

What I find tricky is how to split it up into MRs. I can create a MR that deletes the shapetracker right away, but a gradual refactor might be wiser.