@roelofvandijk 2023-09-10
Consolidated the reshape functions,

```
  0 24079.56 ms run, 24079.43 ms python,    0.14 ms CL, 1187.23 loss, 0.002563 LR, 0.84 GB used,     19.71 GFLOPS
  1 7208.32 ms run, 7208.19 ms python,    0.12 ms CL, 1187.59 loss, 0.002577 LR, 4.39 GB used,     64.97 GFLOPS
  2  217.81 ms run,    5.92 ms python,  211.88 ms CL, 2760.15 loss, 0.001741 LR, 4.39 GB used,   2150.18 GFLOPS
  3  224.96 ms run,   11.24 ms python,  213.72 ms CL, 2844.72 loss, 0.000906 LR, 4.39 GB used,   2081.85 GFLOPS
  4  221.21 ms run,    7.85 ms python,  213.36 ms CL, 2118.59 loss, 0.000070 LR, 4.39 GB used,   2117.06 GFLOPS
```

That's 30%.